# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Science Companion API",
    "description": "长期科学学习与表达伙伴 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scope/context": {
      "get": {
        "tags": [
          "scope"
        ],
        "summary": "Get Scope Context",
        "description": "Return the compiled scope envelope and RLS context for the current request.\n\nThis endpoint is read-only and is used by storage, index and cache layers to\nalign their own scope checks with the API's authorization interpreter.",
        "operationId": "get_scope_context_scope_context_get",
        "parameters": [
          {
            "name": "science_companion_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Science Companion Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scope Context Scope Context Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 401 Get Scope Context Scope Context Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scope/validate-task": {
      "post": {
        "tags": [
          "scope"
        ],
        "summary": "Validate Background Task",
        "description": "Validate that a background task carries all required scope fields.\n\nMissing subject, object domain, authorization version or key epoch causes a\ndeterministic failure closure rather than running with guessed context.",
        "operationId": "validate_background_task_scope_validate_task_post",
        "parameters": [
          {
            "name": "science_companion_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Science Companion Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackgroundTaskEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Background Task Scope Validate Task Post"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 401 Validate Background Task Scope Validate Task Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 403 Validate Background Task Scope Validate Task Post"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 422 Validate Background Task Scope Validate Task Post"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/scope/report-violation": {
      "post": {
        "tags": [
          "scope"
        ],
        "summary": "Report Scope Violation",
        "description": "Accept an audit-safe scope violation report.\n\nThe report intentionally does not include private content or keys.",
        "operationId": "report_scope_violation_scope_report_violation_post",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ScopeAction"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reason"
            }
          },
          {
            "name": "science_companion_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Science Companion Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeEnvelope"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeViolationReport"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 401 Report Scope Violation Scope Report Violation Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response 422 Report Scope Violation Scope Report Violation Post"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "summary": "Me",
        "description": "Protected demo route: returns the resolved subject.\n\nT003 uses this route to prove that API endpoints share the same\nauthentication guard as the Web frontend.",
        "operationId": "me_me_get",
        "parameters": [
          {
            "name": "science_companion_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Science Companion Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Me Me Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/_test/recovery-token": {
      "get": {
        "summary": "Test Recovery Token",
        "description": "Test-only endpoint to retrieve a recovery token without email delivery.\n\nThis endpoint is prefixed with `/_test/` and is only safe because the\nT003 identity service is in-memory. It must not be exposed in production.",
        "operationId": "test_recovery_token__test_recovery_token_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Recovery Token  Test Recovery Token Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/_test/runs/{run_id}/advance": {
      "post": {
        "summary": "Test Advance Run",
        "description": "Test-only endpoint to deterministically advance a run by one node.",
        "operationId": "test_advance_run__test_runs__run_id__advance_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "science_companion_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Science Companion Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunProjection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BackgroundTaskEnvelope": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "Stable background task identifier."
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "description": "Logical task type, e.g. ingestion, index."
          },
          "subject": {
            "$ref": "#/components/schemas/SubjectContext",
            "description": "Subject that authorized the task."
          },
          "object_domain": {
            "$ref": "#/components/schemas/ObjectDomain",
            "description": "Domain in which the task runs."
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id",
            "description": "Project scope when the task belongs to a project."
          },
          "authorization_version": {
            "type": "string",
            "title": "Authorization Version",
            "description": "Authorization policy version at task creation."
          },
          "key_epoch": {
            "type": "string",
            "title": "Key Epoch",
            "description": "Key epoch under which task secrets are bound."
          },
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "description": "Declared purpose of the task."
          },
          "object_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Object Refs",
            "description": "Object identifiers the task is authorized to touch."
          },
          "memory_slice_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Memory Slice Refs",
            "description": "Memory slice identifiers compiled for the task."
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "task_type",
          "subject",
          "object_domain",
          "authorization_version",
          "key_epoch",
          "purpose"
        ],
        "title": "BackgroundTaskEnvelope",
        "description": "Required context for every background task.\n\nA background task must carry a subject, object domain, project, authorization\nversion and key epoch. Missing any of these causes deterministic failure\nclosure rather than running with ambient or guessed context."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunProjection": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Stable run identifier."
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "description": "Project within which the run is scoped."
          },
          "workflow_name": {
            "type": "string",
            "title": "Workflow Name",
            "description": "Compiled workflow template name."
          },
          "workflow_version": {
            "type": "string",
            "title": "Workflow Version",
            "description": "Compiled workflow template version."
          },
          "run_status": {
            "$ref": "#/components/schemas/WorkflowRunStatus",
            "description": "Workflow execution state."
          },
          "artifact_trust_status": {
            "$ref": "#/components/schemas/ArtifactTrustStatus",
            "description": "Scientific trust state of the run's primary artifact."
          },
          "publish_eligible": {
            "type": "boolean",
            "title": "Publish Eligible",
            "description": "Whether all current conditions allow publishing the artifact."
          },
          "objective": {
            "type": "string",
            "title": "Objective",
            "description": "Confirmed task goal."
          },
          "success_criteria": {
            "type": "string",
            "title": "Success Criteria",
            "description": "Confirmed success criteria."
          },
          "risk_statement": {
            "type": "string",
            "title": "Risk Statement",
            "description": "Confirmed risk statement."
          },
          "current_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node Id",
            "description": "Node currently executing or waiting for input."
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeProgress"
            },
            "type": "array",
            "title": "Nodes",
            "description": "All nodes in the compiled workflow."
          },
          "human_todos": {
            "items": {
              "$ref": "#/components/schemas/HumanTodoItem"
            },
            "type": "array",
            "title": "Human Todos",
            "description": "Open and resolved human decisions for the run."
          },
          "run_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Started At"
          },
          "run_ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Ended At"
          },
          "cancel_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Reason"
          },
          "context_envelope": {
            "$ref": "#/components/schemas/RunContextEnvelope",
            "description": "Immutable execution context for refresh and audit."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "project_id",
          "workflow_name",
          "workflow_version",
          "run_status",
          "artifact_trust_status",
          "publish_eligible",
          "objective",
          "success_criteria",
          "risk_statement",
          "context_envelope"
        ],
        "title": "RunProjection",
        "description": "Task-stage projection of a workflow run.\n\nIt separates workflow execution status from artifact trust status and\npublish eligibility, so that users do not confuse \"finished running\" with\n\"scientifically approved\" or \"ready to publish\"."
      },
      "ScopeAction": {
        "type": "string",
        "enum": [
          "create",
          "read",
          "update",
          "delete",
          "share",
          "execute"
        ],
        "title": "ScopeAction",
        "description": "Actions that scope isolation can authorize or deny."
      },
      "ScopeEnvelope": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id",
            "description": "Authenticated account identifier."
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id",
            "description": "Institution/organization tenant when applicable."
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id",
            "description": "Project within which the operation is scoped."
          },
          "object_domain": {
            "$ref": "#/components/schemas/ObjectDomain",
            "description": "Authority domain that owns the target objects.",
            "default": "personal_vault"
          },
          "authorization_version": {
            "type": "string",
            "title": "Authorization Version",
            "description": "Authorization policy version at compile time.",
            "default": "authz-1.0"
          },
          "key_epoch": {
            "type": "string",
            "title": "Key Epoch",
            "description": "Key epoch under which secrets and capsules are bound.",
            "default": "epoch-0"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "description": "Declared purpose for the operation.",
            "default": "general"
          },
          "requested_object_refs": {
            "items": {
              "$ref": "#/components/schemas/ObjectRef"
            },
            "type": "array",
            "title": "Requested Object Refs",
            "description": "Object references the operation requests to touch."
          },
          "requested_vault_refs": {
            "items": {
              "$ref": "#/components/schemas/VaultObjectRef"
            },
            "type": "array",
            "title": "Requested Vault Refs",
            "description": "Vault object references the operation requests to touch."
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "ScopeEnvelope",
        "description": "Immutable scope snapshot carried by requests, runs, cache keys and RLS.\n\nThe envelope binds an operation to a subject, tenant, project, object domain,\nauthorization version and key epoch. Services must compile it before touching\nstorage, cache, index or model context."
      },
      "ScopeViolationReport": {
        "properties": {
          "actor_account_id": {
            "type": "string",
            "title": "Actor Account Id",
            "description": "Account that attempted the action."
          },
          "actor_session_id": {
            "type": "string",
            "title": "Actor Session Id",
            "description": "Session that attempted the action."
          },
          "action": {
            "$ref": "#/components/schemas/ScopeAction",
            "description": "Action that was denied."
          },
          "requested_scope": {
            "$ref": "#/components/schemas/ScopeEnvelope",
            "description": "Scope that was denied."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Human-readable, non-leaking reason."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At",
            "description": "When the violation was detected."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Opaque detail safe for logging; must not expose internal state."
          }
        },
        "type": "object",
        "required": [
          "actor_account_id",
          "actor_session_id",
          "action",
          "requested_scope",
          "reason",
          "occurred_at"
        ],
        "title": "ScopeViolationReport",
        "description": "Audit-safe report when a scope check fails.\n\nThe report intentionally does not include private content, full prompts or\nkeys; it records only the subject, requested scope, reason and timestamp."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}